+++ src/settings/include/settings/general_settings.hpp
@@ -74,6 +74,8 @@ namespace settings
         static constexpr const char* DISMISSED_UPDATE_VERSION_DESCRIPTION =
             "The version of the update that the user has chosen not to be "
             "reminded about.";
+        /// dismissed update version default value
+        static constexpr const char* DISMISSED_UPDATE_VERSION_DEFAULT = "";
     };
 
     /**